requirements.txt
blis                     0.7.4
cachetools               4.2.2
catalogue                2.0.4
certifi                  2020.12.5
cffi                     1.14.5
chardet                  4.0.0
click                    7.1.2
colorama                 0.4.4
cymem                    2.0.5
decorator                4.4.2
google-api-core          1.26.3
google-auth              1.30.0
google-cloud-core        1.6.0
google-cloud-speech      2.4.0
google-cloud-storage     1.38.0
google-crc32c            1.1.2
google-resumable-media   1.2.0
googleapis-common-protos 1.53.0
grpcio                   1.38.0
h5py                     3.2.1
idna                     2.10
imageio                  2.9.0
imageio-ffmpeg           0.4.3
Jinja2                   3.0.1
joblib                   1.0.1
Keras                    2.4.3
libcst                   0.3.18
MarkupSafe               2.0.1
moviepy                  1.0.3
murmurhash               1.0.5
mypy-extensions          0.4.3
nltk                     3.6.2
numpy                    1.20.3
oauthlib                 3.1.0
packaging                20.9
pandas                   1.2.4
pathy                    0.5.2
Pillow                   8.2.0
pip                      21.1.2
preshed                  3.0.5
proglog                  0.1.9
proto-plus               1.18.1
protobuf                 3.16.0
pyasn1                   0.4.8
pyasn1-modules           0.2.8
pycparser                2.20
pydantic                 1.7.4
pyparsing                2.4.7
PySocks                  1.7.1
python-dateutil          2.8.1
pytz                     2021.1
PyYAML                   5.4.1
regex                    2021.4.4
requests                 2.25.1
requests-oauthlib        1.3.0
rsa                      4.7.2
scikit-learn             0.24.2
scipy                    1.6.3
setuptools               47.1.0
six                      1.16.0
smart-open               3.0.0
spacy                    3.0.6
spacy-legacy             3.0.5
srsly                    2.4.1
thinc                    8.0.3
threadpoolctl            2.1.0
tqdm                     4.60.0
tweepy                   3.10.0
twitter-media-downloader 1.0.0
typer                    0.3.2
typing-extensions        3.10.0.0
typing-inspect           0.6.0
urllib3                  1.26.4
wasabi                   0.8.2
wget                     3.2